+++ main/urls.py
@@ -1,12 +1,13 @@
 from django.urls import path, include
 from users.urls import auth_patterns, friends_patterns, settings_patterns
+from engagement.urls import leaderboard_patterns, achievement_patterns
 from . import views
 
 # URL patterns for the main application
 urlpatterns = [
     path("", views.home, name="home"),  # Home page
     # path("qr", views.generate_qr, name="qr"),  # QR code generation page
-    path("leaderboard", views.leaderboard, name="leaderboard"),  # Leaderboard page
+
     path("allchallenges", views.my_challenges, name="allchallenges"),  # All challenges page
     # path("createchallenge", views.createChallenge, name="createchallenge"),
     # path("save", views.save_image,name="save" ),
@@ -16,13 +17,15 @@ urlpatterns = [
     path("friends/", include(friends_patterns)),
     path("settings/", include(settings_patterns)),
 
+    path("leaderboard/", include(leaderboard_patterns)),
+    path("achievement/", include(achievement_patterns)),
+
     path("events", views.events, name="events"),  # Events page
     path("market/", views.market_view, name="market"),  # Market page
     # path("addChallenge", views.add_challenge),  # Add challenge endpoint
     # path("api/v1/addChallenge", views.add_challenge),  # API endpoint to add challenge
     path("api/v1/removeChallenge", views.remove_challenge),  # API endpoint to remove challenge
     path("api/v1/removeTask", views.remove_task),  # API endpoint to remove task
-    path("api/v1/leaderboard", views.get_leaderboard),  # API endpoint to get leaderboard
     path('incrementProgress/<int:event_id>/', views.incrementProgress, name='incrementProgress'),  # Increment progress endpoint
     path("market/api/add_purchased_plant", views.add_purchased_plant, name="add_purchased_plant"),  # Add purchased plant endpoint
     path('delete_event/<int:event_id>/', views.delete_event, name='delete_event'),
