+++ users/models.py
@@ -1,8 +1,6 @@
 from django.db import models
 from django.utils.timezone import now
 from django.contrib.auth.models import AbstractUser, Group, BaseUserManager
-from django.db.models.signals import post_save
-from django.dispatch import receiver
 
 class CustomUserManager(BaseUserManager):
     """Custom manager to prevent issues with swapped user models."""
@@ -110,18 +108,6 @@ class CustomUser(AbstractUser):
         ).values_list("user1", flat=True)
 
         return CustomUser.objects.filter(id__in=request_ids)
-
-class UserStats(models.Model):
-    """Model to store user stats like leaves and points."""
-    user = models.OneToOneField(CustomUser, on_delete=models.CASCADE, related_name="stats")
-    leaves = models.IntegerField(default=0)
-    points = models.IntegerField(default=0)
-    class Meta:
-        verbose_name = "User Stats"
-        verbose_name_plural = "User Stats"
-
-    def __str__(self):
-        return f"| {self.user.username} | {self.leaves} Leaves Remaining | {self.points} Total Points |"
 
 class Friendship(models.Model):
     """Model to track friendships explicitly, replacing the ManyToManyField."""
